+++ packages/data-types/src/lib/enums/weather.enum.ts
@@ -1,8 +1,0 @@
-export enum Weather {
-    SUNNY = 'Sunny',
-    RAIN = 'Rain',
-    STORM = 'Storm',
-    WINDY = 'Windy',
-    SNOW = 'Snow',
-    BLIZZARD = 'Blizzard'
-}