+++ packages/guide/src/app/journal/components/base-journal-page/base-journal-page.component.ts
@@ -1,5 +1,5 @@
-import { ChangeDetectorRef, Component, inject, ViewChild } from '@angular/core';
-import { BaseCrop, JournalOrder, MinimalItem, Season, UiIcon } from '@ci/data-types';
+import { ChangeDetectorRef, Component, inject, OnDestroy, ViewChild } from '@angular/core';
+import { BaseCrop, JournalOrder, MinimalItem, UiIcon } from '@ci/data-types';
 import { combineLatest, map, Observable, of, startWith } from 'rxjs';
 import { MediaMatcher } from '@angular/cdk/layout';
 import { FormGroup } from '@angular/forms';
@@ -17,7 +17,7 @@ export interface BaseJournalPageComponent<D> {
 })
 export class BaseJournalPageComponent<D extends ({
     item: MinimalItem
-} | MinimalItem)> extends BaseTabbedSelectableContainerComponent<D> {
+} | MinimalItem)> extends BaseTabbedSelectableContainerComponent<D> implements OnDestroy {
 
     @ViewChild(MatTabGroup) matTabGroup?: MatTabGroup