+++ packages/guide/src/app/journal/components/tables/found-table/found-table.component.ts
@@ -1,14 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
     selector: 'app-found-table',
     templateUrl: './found-table.component.html',
-    styleUrls: ['./found-table.component.scss'],
 })
-export class FoundTableComponent implements OnInit {
-    constructor() {
-    }
-
-    ngOnInit(): void {
-    }
+export class FoundTableComponent {
 }