+++ packages/guide/src/app/shared/pipes/is-tag-based-item.pipe.ts
@@ -1,11 +1,12 @@
-import { Pipe } from '@angular/core';
+import { Pipe, PipeTransform } from '@angular/core';
 import { TagBasedItem } from "@ci/data-types";
 import { IsMinimalTagBasedItemPipe } from "./is-minimal-tag-based-item.pipe";
 
 @Pipe({
     name: 'isTagBasedItem',
 })
-export class IsTagBasedItemPipe extends IsMinimalTagBasedItemPipe {
+export class IsTagBasedItemPipe extends IsMinimalTagBasedItemPipe implements PipeTransform {
+
     override transform(value: any): value is TagBasedItem {
         return super.transform(value) && 'items' in value;
     }