+++ packages/guide/src/app/shared/services/base-db.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { BaseDbService } from './base-db.service';
+
+describe('BaseDbService', () => {
+  let service: BaseDbService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(BaseDbService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});