+++ packages/guide/src/CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added description, birthday and datable status to NPC overview
 - Added enchantment points to found items in journal
 - Added fish pattern and difficulty to journal and details
+- Added permanently visible label for inputs to show what they represent
+- Added museum fish to My Coral Guide dashboard
 
 #### Changed